refactor: 패키지 정의 및 분리
