+++ src/main/kotlin/racing/domain/Car.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.domain
 
 class Car(
     val id: Int,
