+++ src/main/kotlin/racing/domain/InputView.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.domain
 
 interface InputView {
     fun askTryCount()
