+++ src/main/kotlin/racing/domain/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.domain
 
 fun interface NumberGenerator {
     fun generate(): Int
