+++ src/main/kotlin/racing/domain/ResultView.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.domain
 
 fun interface ResultView {
     fun showResult(result: RaceResult)
