+++ src/main/kotlin/racing/domain/Vehicle.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.domain
 
 fun interface Vehicle {
     fun move()
