+++ src/main/kotlin/racing/presentation/request/RaceRequest.kt
@@ -1,0 +1,6 @@
+package racing.presentation.request
+
+data class RaceRequest(
+    val names: List<String>,
+    val tryCount: Int,
+)
