+++ src/test/kotlin/racing/CarFactoryTest.kt
@@ -2,6 +2,8 @@ package racing
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
+import racing.domain.Car
+import racing.domain.CarFactory
 
 class CarFactoryTest : StringSpec({
     val sut = CarFactory
