+++ src/test/kotlin/racing/FixedNumberGenerator.kt
@@ -1,5 +1,7 @@
 package racing
 
+import racing.domain.NumberGenerator
+
 class FixedNumberGenerator(
     private val number: Int,
 ) : NumberGenerator {
