+++ src/test/kotlin/racing/RandomCountGeneratorTest.kt
@@ -2,6 +2,7 @@ package racing
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.ints.shouldBeBetween
+import racing.infrastructure.RandomNumberGenerator
 
 class RandomCountGeneratorTest : StringSpec({
     "0 ~ 9 사이의 임의의 숫자를 생성할 수 있다" {
